#33 - Improving the feedbacks and validations of the API Permissions.
